LibWeb: Split LinearGradientStyleValue out of StyleValue.{h,cpp}
