+++ Userland/Libraries/LibWeb/Painting/GradientPainting.cpp
@@ -8,6 +8,7 @@
 #include <LibGfx/Gradients.h>
 #include <LibWeb/CSS/StyleValue.h>
 #include <LibWeb/CSS/StyleValues/ConicGradientStyleValue.h>
+#include <LibWeb/CSS/StyleValues/LinearGradientStyleValue.h>
 #include <LibWeb/Painting/GradientPainting.h>
 
 namespace Web::Painting {
